+++ HISTORY.md
@@ -2,7 +2,7 @@
    
 * 20170420, V0.9.4
     * Added periodic timer support which provides better accuracy than
-      setInterval().
+      setInterval()
     * Updated dependencies
     * Updated copyright notice